Add pet data for Chitterspine Grotto
It's a cave off of Nazjatar, but has its own map and two wild pets that spawn there, one of which is nowhere else.

Two pets there:
* Hissing Chitterspine (unique)
* Chitterspine Skitterling (also elsewhere)

Added a location for Hissing Chitterspine for the entrance of the cave as well, so that people going around Nazjatar can find it more easily.